patch 16.06.18

+++ Assets/Scripts/Player/Essence.cs
@@ -20,5 +20,5 @@ public class Essence : ScriptableObject {
 	public int defaultDamage;
 	public int defaultHealth;
 	public int defaultArmor;
-	public int attackSpeed;
+	public float attackSpeed;
 }
